fix: too many search results
When more than 1,000 repositories are found based on the specified search criteria, the program will warn the user and return results only from the first 1,000 repositories found. Previously, an error would be thrown when more than the first 1,000 results tried to be retrieved, as this is a GitHub API limitation.